The interface for passing values supports all java basic types

+++ src/java/test/src/androidTest/java/com/tencent/wcdbtest/crud/StatementOperationTest.java
@@ -121,7 +121,7 @@ public class StatementOperationTest extends ValueCRUDTestCase {
         PreparedStatement prepareSelect = handle.getOrCreatePreparedStatement(select);
         prepareSelect.step();
         assertFalse(prepareSelect.isDone());
-        assertEquals(prepareSelect.getInteger(0), value);
+        assertEquals(prepareSelect.getLong(0), value);
     }
 
     @Test
